requirements.txt
openai>=1.0
langchain>=0.3
langchain-openai>=0.3
pandas>=2.0
python-dotenv>=1.0
